+++ back-end/src/main/kotlin/com/apexdevs/backend/persistence/exception/SynthesisFlagException.kt
@@ -4,8 +4,8 @@ import nl.uu.cs.ape.sat.models.enums.SynthesisFlag
 
 class SynthesisFlagException(val from: Any, val flag: SynthesisFlag) : RuntimeException(flag.message) {
     /**
-     * Overrides some of APE's SynthesisFlag messages with user friendly messages which can be shown on the front-end.
-     * @return A user friendly message describing the reason the synthesis was interrupted.
+     * Overrides some of APE's SynthesisFlag messages with user-friendly messages which can be shown on the front-end.
+     * @return A user-friendly message describing the reason the synthesis was interrupted.
      */
     fun getFriendlyMessage(): String {
         return when (flag) {